refactor(crypto): Migrate human hasher to util
