+++ packages/common/util/package.json
@@ -17,12 +17,13 @@
     "test": "toolchain test"
   },
   "dependencies": {
-    "@dxos/crypto": "workspace:*",
     "@dxos/debug": "workspace:*",
     "@dxos/protocols": "workspace:*"
   },
   "devDependencies": {
+    "@dxos/crypto": "workspace:*",
     "@dxos/eslint-plugin": "~1.0.34",
+    "@dxos/protocols": "workspace:*",
     "@dxos/protocols-toolchain": "workspace:*",
     "@types/jest": "^26.0.7",
     "@types/node": "^16.11.27",
