+++ packages/common/util/src/human-hash.test.ts
@@ -1,0 +1,18 @@
+//
+// Copyright 2020 DXOS.org
+//
+
+// DXOS testing browser.
+
+import { createKeyPair, createId } from '@dxos/crypto';
+import { PublicKey } from '@dxos/protocols';
+
+import { humanize } from './human-hash';
+
+test('Hashing', () => {
+  const { publicKey, secretKey } = createKeyPair();
+
+  expect(humanize(publicKey)).not.toEqual(humanize(secretKey));
+  expect(humanize(publicKey)).toEqual(humanize(PublicKey.from(publicKey)));
+  expect(humanize(createId())).toBeDefined();
+});
