+++ packages/demos/tutorials-tasks-app/package.json
@@ -26,7 +26,6 @@
   "dependencies": {
     "@dxos/client": "workspace:*",
     "@dxos/config": "workspace:*",
-    "@dxos/crypto": "workspace:*",
     "@dxos/echo-db": "workspace:*",
     "@dxos/object-model": "workspace:*",
     "@dxos/react-client": "workspace:*",
