+++ packages/devtools/devtools-mesh/package.json
@@ -12,7 +12,6 @@
     "lint": "toolchain lint"
   },
   "dependencies": {
-    "@dxos/crypto": "workspace:*",
     "@dxos/debug": "workspace:*",
     "@dxos/gem-core": "~2.2.1-beta.11",
     "@dxos/gem-spore": "~2.2.1-beta.11",
@@ -20,6 +19,7 @@
     "@dxos/protocol-plugin-presence": "workspace:*",
     "@dxos/react-components": "workspace:*",
     "@dxos/react-toolkit": "workspace:*",
+    "@dxos/util": "workspace:*",
     "@emotion/css": "^11.7.1",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
