+++ packages/mesh/protocol-plugin-rpc/package.json
@@ -14,7 +14,6 @@
   },
   "dependencies": {
     "@dxos/async": "workspace:*",
-    "@dxos/crypto": "workspace:*",
     "@dxos/mesh-protocol": "workspace:*",
     "@dxos/rpc": "workspace:*",
     "@dxos/util": "workspace:*"
