+++ packages/mesh/protocol/src/utils.ts
@@ -4,14 +4,12 @@
 
 import assert from 'assert';
 
-import { HumanHasher } from '@dxos/crypto';
-
-const hasher = new HumanHasher();
+import { humanize } from '@dxos/util';
 
 export const keyToHuman = (key: Buffer, prefix?: string) => {
   assert(Buffer.isBuffer(key));
 
-  const name = hasher.humanize(key.toString('hex'));
+  const name = humanize(key.toString('hex'));
   if (prefix) {
     return `${prefix}(${name})`;
   }
