+++ packages/mesh/signal/package.json
@@ -33,7 +33,6 @@
   "dependencies": {
     "@azure/abort-controller": "^1.0.1",
     "@dxos/broadcast": "workspace:*",
-    "@dxos/crypto": "workspace:*",
     "@hyperswarm/dht": "^4.0.1",
     "buffer-json-encoding": "^1.0.2",
     "debug": "^4.3.3",
