+++ packages/sdk/react-client/package.json
@@ -32,7 +32,6 @@
     "use-subscription": "^1.8.0"
   },
   "devDependencies": {
-    "@dxos/crypto": "workspace:*",
     "@dxos/esbuild-plugins": "~2.28.10",
     "@dxos/eslint-plugin": "~1.0.34",
     "@dxos/protocols-toolchain": "workspace:*",
