+++ packages/sdk/react-components/package.json
@@ -20,8 +20,8 @@
   },
   "dependencies": {
     "@dxos/async": "workspace:*",
-    "@dxos/crypto": "workspace:*",
     "@dxos/debug": "workspace:*",
+    "@dxos/util": "workspace:*",
     "@emotion/css": "^11.7.1",
     "debug": "^4.3.3",
     "lodash.defaultsdeep": "^4.6.1",
