+++ packages/sdk/react-components/src/MemberList/MemberAvatar.tsx
@@ -11,8 +11,8 @@ import {
 import { Avatar, Tooltip, colors, useTheme } from '@mui/material';
 
 import { PartyMember } from '@dxos/client';
-import { humanize } from '@dxos/crypto';
 import { PublicKeyLike } from '@dxos/protocols';
+import { humanize } from '@dxos/util';
 
 const { red, pink, deepPurple, deepOrange, indigo, blue, cyan, teal, green, amber } = colors;
 const depth = 500;
